Do not escape HTML in package.json output.
The Go JSON encoder escapes characters such as < and > in its output by
default, resulting in "author" strings like "foo <foo@bar.io>" being
read from package.json correctly, but on any writes gx does to the file
it is written as "foo \u003cfoo@bar.io\u003e".

This was done as a safety precaution to keep HTML from being embedded in
JSON payloads, but this does not seem to apply in the case of a
package.json file.

This patch modifies the package.json writing logic to not escape HTML.