tests: fix PATH precedence, and clarify the feedback
Sharness adds ./tests/.. to PATH. If you had a gx executable
in ./ (e.g. as a result of `go build`), it would shadow the
fresh test executable, and the test would be bogus.

This patch fixes the precedence issue by adding ./tests/bin last.

It also clarifies the error returned if the test ipfs executable
is shadowed, while we're at it.