Split `InstallDeps`
Split the `InstallDeps` function in two functions which represent clear separate
processes: fetching packages (`fetchDependencies`) and running the `post-
install` hook on them (`dependenciesPostInstall`).

The `fetchDependencies` function is the previous `installDeps` function (now
renamed) without the `maybeRunPostInstall` call, which was encapsulated in its
own `dependenciesPostInstall` function that replicates the same DFS traversal of
the dependency DAG to run the hook.

This commit permits a more clear understanding of the process of installing
dependencies which clears the road for the unification of the fetching functions
into one.

There is a count discrepancy bug in the `ProgMeter` introduced in this commit
but it will be fixed in the next commit when `fetchDependencies` is refactored
to eliminate the recursive implementation.