gxutil: unify dependency fetching in `InstallDeps`
The previous implementation spawned multiple goroutines to fetch packages but
for the dependencies of one package at a time. This meant that the last
goroutine that took longer to finish (probably because it was fetching the
biggest dependency) held the start of more goroutines to fetch the dependencies
of the packages already fetched.

Unify all the dependencies in a single queue (abstracted in the new
`DependencyQueue` structure) so as soon as new dependencies of a fetched package
are available new goroutines can be spawned to fetch them.

Raise the number of parallel goroutines fetching packages (now `maxGoroutines`)
from 2 to 20 now that this bigger number can be leveraged from this unification.

Remove the retry logic around `GetPackageTo` (as that function already contains
code with equivalent functionality) to simplify the fetching part of the
function.